refactored compute script releases to edge scripting releases
