--- src/Model/API/Base/Compute/ListComputeScriptReleases.php
@@ -1,40 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace ToshY\BunnyNet\Model\API\Base\Compute;
-
-use ToshY\BunnyNet\Enum\Header;
-use ToshY\BunnyNet\Enum\Method;
-use ToshY\BunnyNet\Enum\Type;
-use ToshY\BunnyNet\Model\AbstractParameter;
-use ToshY\BunnyNet\Model\EndpointInterface;
-use ToshY\BunnyNet\Model\EndpointQueryInterface;
-
-class ListComputeScriptReleases implements EndpointInterface, EndpointQueryInterface
-{
-    public function getMethod(): Method
-    {
-        return Method::GET;
-    }
-
-    public function getPath(): string
-    {
-        return 'compute/script/%d/releases';
-    }
-
-    public function getHeaders(): array
-    {
-        return [
-            Header::ACCEPT_JSON,
-        ];
-    }
-
-    public function getQuery(): array
-    {
-        return [
-            new AbstractParameter(name: 'page', type: Type::INT_TYPE),
-            new AbstractParameter(name: 'perPage', type: Type::INT_TYPE),
-        ];
-    }
-}
