+++ src/Model/API/EdgeScripting/Release/GetActiveReleases.php
@@ -1,0 +1,29 @@
+<?php
+
+declare(strict_types=1);
+
+namespace ToshY\BunnyNet\Model\API\EdgeScripting\Release;
+
+use ToshY\BunnyNet\Enum\Header;
+use ToshY\BunnyNet\Enum\Method;
+use ToshY\BunnyNet\Model\EndpointInterface;
+
+class GetActiveReleases implements EndpointInterface
+{
+    public function getMethod(): Method
+    {
+        return Method::GET;
+    }
+
+    public function getPath(): string
+    {
+        return 'compute/script/%d/releases/active';
+    }
+
+    public function getHeaders(): array
+    {
+        return [
+            Header::ACCEPT_JSON,
+        ];
+    }
+}
