+++ src/Model/API/EdgeScripting/Release/GetReleases.php
@@ -2,39 +2,39 @@
 
 declare(strict_types=1);
 
-namespace ToshY\BunnyNet\Model\API\Base\Compute;
+namespace ToshY\BunnyNet\Model\API\EdgeScripting\Release;
 
 use ToshY\BunnyNet\Enum\Header;
 use ToshY\BunnyNet\Enum\Method;
 use ToshY\BunnyNet\Enum\Type;
 use ToshY\BunnyNet\Model\AbstractParameter;
-use ToshY\BunnyNet\Model\EndpointBodyInterface;
 use ToshY\BunnyNet\Model\EndpointInterface;
+use ToshY\BunnyNet\Model\EndpointQueryInterface;
 
-class PublishComputeScriptByPathParameter implements EndpointInterface, EndpointBodyInterface
+class GetReleases implements EndpointInterface, EndpointQueryInterface
 {
     public function getMethod(): Method
     {
-        return Method::POST;
+        return Method::GET;
     }
 
     public function getPath(): string
     {
-        return 'compute/script/%d/publish/%s';
+        return 'compute/script/%d/releases';
     }
 
     public function getHeaders(): array
     {
         return [
             Header::ACCEPT_JSON,
-            Header::CONTENT_TYPE_JSON,
         ];
     }
 
-    public function getBody(): array
+    public function getQuery(): array
     {
         return [
-            new AbstractParameter(name: 'Note', type: Type::STRING_TYPE),
+            new AbstractParameter(name: 'page', type: Type::INT_TYPE),
+            new AbstractParameter(name: 'perPage', type: Type::INT_TYPE),
         ];
     }
 }
