+++ src/Model/API/EdgeScripting/Release/PublishRelease.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace ToshY\BunnyNet\Model\API\Base\Compute;
+namespace ToshY\BunnyNet\Model\API\EdgeScripting\Release;
 
 use ToshY\BunnyNet\Enum\Header;
 use ToshY\BunnyNet\Enum\Method;
@@ -10,9 +10,8 @@ use ToshY\BunnyNet\Enum\Type;
 use ToshY\BunnyNet\Model\AbstractParameter;
 use ToshY\BunnyNet\Model\EndpointBodyInterface;
 use ToshY\BunnyNet\Model\EndpointInterface;
-use ToshY\BunnyNet\Model\EndpointQueryInterface;
 
-class PublishComputeScript implements EndpointInterface, EndpointQueryInterface, EndpointBodyInterface
+class PublishRelease implements EndpointInterface, EndpointBodyInterface
 {
     public function getMethod(): Method
     {
@@ -27,15 +26,7 @@ class PublishComputeScript implements EndpointInterface, EndpointQueryInterface,
     public function getHeaders(): array
     {
         return [
-            Header::ACCEPT_JSON,
             Header::CONTENT_TYPE_JSON,
-        ];
-    }
-
-    public function getQuery(): array
-    {
-        return [
-            new AbstractParameter(name: 'uuid', type: Type::STRING_TYPE, required: true),
         ];
     }
