+++ src/Model/API/EdgeScripting/Release/PublishReleaseByPathParameter.php
@@ -1,0 +1,39 @@
+<?php
+
+declare(strict_types=1);
+
+namespace ToshY\BunnyNet\Model\API\EdgeScripting\Release;
+
+use ToshY\BunnyNet\Enum\Header;
+use ToshY\BunnyNet\Enum\Method;
+use ToshY\BunnyNet\Enum\Type;
+use ToshY\BunnyNet\Model\AbstractParameter;
+use ToshY\BunnyNet\Model\EndpointBodyInterface;
+use ToshY\BunnyNet\Model\EndpointInterface;
+
+class PublishReleaseByPathParameter implements EndpointInterface, EndpointBodyInterface
+{
+    public function getMethod(): Method
+    {
+        return Method::POST;
+    }
+
+    public function getPath(): string
+    {
+        return 'compute/script/%d/publish/%s';
+    }
+
+    public function getHeaders(): array
+    {
+        return [
+            Header::CONTENT_TYPE_JSON,
+        ];
+    }
+
+    public function getBody(): array
+    {
+        return [
+            new AbstractParameter(name: 'Note', type: Type::STRING_TYPE),
+        ];
+    }
+}
